[PPT-10.2] Task 3 fixups: guard geometry observer + move Redis fixture to conftest
- _on_electrode_channel_changed: early-return when _disable_state_messages
  is True, preventing intermediate partial publishes during bulk apply_message_model
- Move session-scoped _mock_redis_for_dock_pane_import fixture from
  test_geometry_publish.py to device_viewer/tests/conftest.py so it applies
  regardless of test collection order; pre-delete cached modules before
  patch.dict so teardown correctly restores the absent state
